# 📊 Make a simpler merge.