Replace the complex tournament tree merge by a simple linear merge.
After:
```
❯ ARBENV_NUM_THREADS=1 hyperfine --warmup=2 'bin/brunel -G 10 -m 1000 -n 1000 -t 200'
Benchmark 1: bin/brunel -G 10 -m 1000 -n 1000 -t 200
  Time (mean ± σ):      2.083 s ±  0.024 s    [User: 2.041 s, System: 0.038 s]
  Range (min … max):    2.060 s …  2.140 s    10 runs
```
Before
```
❯ ARBENV_NUM_THREADS=1 hyperfine --warmup=2 'bin/brunel -G 10 -m 1000 -n 1000 -t 200'
Benchmark 1: bin/brunel -G 10 -m 1000 -n 1000 -t 200
  Time (mean ± σ):      2.607 s ±  0.007 s    [User: 2.564 s, System: 0.041 s]
  Range (min … max):    2.593 s …  2.617 s    10 runs
```
20% speed-up on an event-heavy setting by removing complex code ¯\_(ツ)_/¯

Dear review, please check on a different machine by toggling `#if 0` in `merge_events.cpp`.