@boeschf 2023-04-03
> `arb::merge_events` is now aware of this and switches from linear to pqueue merge at 20.

where can I find this in the code?