@thorstenhater 2023-04-05
> > `arb::merge_events` is now aware of this and switches from linear to pqueue merge at 20.
> 
> where can I find this in the code?

In `merge_events` is an `if` toggling between `pqueue` and `linear` at 20 queues.