@boeschf 2023-04-05
There are a couple of changes regarding currents (`current_written` etc) in the multicore backend. What about the GPU backend?