@thorstenhater 2023-04-06
Yes, the conditional reset on current write should be propagated. As for zero'ing arrays: cuda_memset is AFAIR not async, while our fill kernel is, so I'll vote to keep it as is.