@thorstenhater 2023-04-17
I am sharding this into individual bits to make debugging and review a bit easier.
- CMake: https://github.com/arbor-sim/arbor/pull/2106